#57 move Elasticsearch7 specific implementation and property constants to its own ".v7" package - the FQN "com.aconex.scrutineer.elasticsearch.v7.ElasticSearchStreamConnector" is part of configuration contract now, the name has to be unique
