+++ elasticsearch7/src/main/java/com/aconex/scrutineer/elasticsearch/v7/TransportAddressParser.java
@@ -1,4 +1,4 @@
-package com.aconex.scrutineer.elasticsearch;
+package com.aconex.scrutineer.elasticsearch.v7;
 
 import java.net.InetAddress;
 import java.net.UnknownHostException;
