+++ elasticsearch7/src/test/java/com/aconex/scrutineer/elasticsearch/v7/TransportAddressParserTest.java
@@ -1,4 +1,4 @@
-package com.aconex.scrutineer.elasticsearch;
+package com.aconex.scrutineer.elasticsearch.v7;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
@@ -11,6 +11,7 @@ import java.net.InetAddress;
 import java.net.UnknownHostException;
 import java.util.List;
 
+import com.aconex.scrutineer.elasticsearch.v7.TransportAddressParser;
 import org.elasticsearch.common.transport.TransportAddress;
 import org.junit.Test;
 import org.junit.runner.RunWith;
