+++ scrutineer/config/elasticsearch7.properties
@@ -1,4 +1,4 @@
-stream.connector.class=com.aconex.scrutineer.elasticsearch.ElasticSearchStreamConnector
+stream.connector.class=com.aconex.scrutineer.elasticsearch.v7.ElasticSearchStreamConnector
 es.cluster.name=myCluster
 es.hosts=localhost:9300
 es.index.name=TestIndex
