+++ scrutineer/src/test/resources/test-elasticsearch7-config.properties
@@ -1,4 +1,4 @@
-stream.connector.class=com.aconex.scrutineer.elasticsearch.ElasticSearchStreamConnector
+stream.connector.class=com.aconex.scrutineer.elasticsearch.v7.ElasticSearchStreamConnector
 es.cluster.name=scrutineerintegrationtest
 es.hosts=localhost:9300
 es.index.name=test
